+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/k8s_attributes.py
@@ -39,9 +39,9 @@ UUIDs as standardized by
 Which states:
 
 > If generated according to one of the mechanisms defined in Rec.
-  ITU-T X.667 | ISO/IEC 9834-8, a UUID is either guaranteed to be
-  different from all other UUIDs generated before 3603 A.D., or is
-  extremely likely to be different (depending on the mechanism chosen).
+> ITU-T X.667 | ISO/IEC 9834-8, a UUID is either guaranteed to be
+> different from all other UUIDs generated before 3603 A.D., or is
+> extremely likely to be different (depending on the mechanism chosen).
 
 Therefore, UIDs between clusters should be extremely unlikely to
 conflict.